moved all scss files to folder -> styles
