--- src/components/Alltitsu.scss
@@ -1,14 +1,0 @@
-@import "../colorvariables.scss";
-
-
-.alltitsuStyling {
-    --digi--navigation-vertical-menu-item--active-indicator-background-color: #41337a;
-    --digi--color--background--inverted-6: #6ea4bf;
-    --digi--navigation-sidebar--wrapper--background-color: #c2efeb;
-    --digi--color--background--inverted-1: #ecfee8;
-    --digi--color--text--inverted: #41337a;
-    --digi--button--text-align--default: center;
-    --digi--button--color--border--primary--default: #c2efeb;
-    --digi--link-external--color--visited: #41337a;
-    --digi--link-external--color--default: #331e36;
-}
