+++ src/components/App.tsx
@@ -1,6 +1,6 @@
-import './App.scss';
+import '../styles/App.scss';
 import { RouterProvider } from 'react-router-dom';
-import { router } from './Router';
+import { router } from '../Router';
 
 export function App() {
   return <RouterProvider router={router}></RouterProvider>;
