+++ src/styles/ContactForm.scss
@@ -1,4 +1,4 @@
-@import '../colorvariables.scss';
+@import './colorvariables.scss';
 
 .contact-btn {
   display: flex;
