+++ src/styles/Footer.scss
@@ -1,4 +1,4 @@
-@import '../colorvariables.scss';
+@import './colorvariables.scss';
 
 .footer-left {
   display: flex;
