+++ src/styles/JobCard.scss
