+++ src/styles/SearchForm.scss
@@ -1,4 +1,4 @@
-@import '../colorvariables.scss';
+@import './colorvariables.scss';
 @import '@digi/arbetsformedlingen/dist/digi-arbetsformedlingen/digi-arbetsformedlingen.css';
 
 .searchForm {
